Extract helper types
