+++ src/Analyzers/Core/Analyzers/Analyzers.projitems
@@ -83,6 +83,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)UseCollectionInitializer\AbstractObjectCreationExpressionAnalyzer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UseCollectionInitializer\AbstractUseCollectionInitializerDiagnosticAnalyzer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UseCollectionInitializer\AbstractUseCollectionInitializerAnalyzer.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UseCollectionInitializer\IUpdateExpressionSyntaxHelper.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UseCollectionInitializer\UpdateExpressionState.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UseCollectionInitializer\UpdateObjectCreationHelpers.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UseCollectionInitializer\UseCollectionInitializerHelpers.cs" />
