+++ src/Analyzers/Core/Analyzers/UseCollectionInitializer/IUpdateExpressionSyntaxHelper.cs
@@ -1,0 +1,17 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Collections.Generic;
+
+namespace Microsoft.CodeAnalysis.UseCollectionInitializer;
+
+internal interface IUpdateExpressionSyntaxHelper<
+    TExpressionSyntax,
+    TStatementSyntax>
+    where TExpressionSyntax : SyntaxNode
+    where TStatementSyntax : SyntaxNode
+{
+    void GetPartsOfForeachStatement(TStatementSyntax statement, out SyntaxToken identifier, out TExpressionSyntax expression, out IEnumerable<TStatementSyntax> statements);
+    void GetPartsOfIfStatement(TStatementSyntax statement, out TExpressionSyntax condition, out IEnumerable<TStatementSyntax> whenTrueStatements, out IEnumerable<TStatementSyntax>? whenFalseStatements);
+}
